Add solution for problem 1238 (BOJ)
1238번: Silver Cow Party

The solution, written in both Rust and C++, used Dijkstra's algorithm,
which I needed to revisit in order to write one.

It was interesting to see that while the algorithm calculates distances
from a certain point to others, it can be misused to calculate the
inverse just by reversing the directions of the graph.