Add solution for problem 236 (LeetCode)
199. Binary Tree Right Side View

It was my first solution for LeetCode that uses breadth-first search.
However, it was not my first time using the algorithm, and was able to
solve the problem with ease.

I used a queue and kept track of nodes for the specific level, which was
something that I have been used to for quite a while.